backend/requirements.txt
pydentic
groq
pymupdf
python-dotenv
supabase
pandas